Prevalence of Diabetes.
Diabetes is one of the most prevalent diseases in our modern world, there are many factor for this. In this analysis made on Jupyter Notebooks, there is a review of this in a given sample of 768 patients, where there were studied 9 factors considered important to the conclusion of this analysis.